1st update

+++ Admin/Enrollment.cs
@@ -14,8 +14,9 @@ namespace Admin
 {
     public partial class Enrollment : Form
     {
-
+        string settedcb = "";
         MySqlConnection conn;
+        string selectedId = "0", selectedstudid, selectedyear, selectedsubid, selecteddesc;
         //============ MY BUILT-IN METHODS =================
         private void databaseConnection()
         {
@@ -27,36 +28,95 @@ namespace Admin
                 string dbase = "database=cruddatabase;";
                 conn = new MySqlConnection(server + uid + pass + dbase);
                 conn.Open();
-                displayEnrollmentTable();
-            }
-            catch (Exception e)
+            }catch (Exception e)
             {
-                MessageBox.Show(e.Message.ToString());
+                MessageBox.Show("Error enrollment form DB connection: " + e.Message.ToString());
             }
 
         }
 
         private void displayEnrollmentTable()
         {
-            string query = "SELECT enrollment_id, stud_id, lname, fname, block, year FROM enrollment_tbl";
+            string query = "SELECT enrollment_id, stud_id, stud_year, sub_id, sub_taken FROM enrollment_tbl";
             MySqlCommand cmd = new MySqlCommand(query, conn);
             MySqlDataReader dr = cmd.ExecuteReader();
             while (dr.Read())
             {
-                enrollment_table.Rows.Add(dr[0].ToString(), dr[1].ToString(), dr[2].ToString(), dr[3].ToString(), dr[4].ToString(), dr[5].ToString());
+                enrollment_table.Rows.Add(dr[0].ToString(), dr[1].ToString(), dr[2].ToString(), dr[3].ToString(), dr[4].ToString());
             }
             dr.Close();
         }
 
 
-        public void deleteSelectedRow(string id)
+        void deleteValue(string id, string subname)
         {
-            string query = "DELETE FROM enrollment_tbl WHERE enrollment_id =" + id + "";
-            MySqlCommand cmd = new MySqlCommand(query, conn);
-            cmd.ExecuteNonQuery();
-            enrollment_table.Refresh();
+            try
+            {
+                DialogResult dr = MessageBox.Show("Do you want to drop this student?\n\nID: " + selectedstudid + "\nSubject Taken: " + subname + "", "Notice!", MessageBoxButtons.OKCancel);
+                if (dr == DialogResult.OK)
+                {
+                    string query = "DELETE FROM enrollment_tbl WHERE enrollment_id =" + id + "";
+                    MySqlCommand cmd = new MySqlCommand(query, conn);
+                    cmd.ExecuteNonQuery();
+                }
+            }
+            catch (Exception e)
+            {
+                MessageBox.Show(e.Message + " on deleteValue function -> Student form");
+                displayEnrollmentTable();
+            }
+
+            displayEnrollmentTable();
+            selectedId = "0";
+
         }
 
+
+        private void displaySearchResult()
+        {
+            enrollment_table.Rows.Clear();
+            try
+            {
+                string query = "SELECT enrollment_id, stud_id, stud_year, sub_id, sub_taken FROM enrollment_tbl";
+                string condition = "";
+
+                if (settedcb == "Student Id" || settedcb == "--Set--")
+                {
+                    condition = "WHERE stud_id LIKE '" + search.Text.ToString() + "%'";
+                }
+                if (settedcb == "Subject Id")
+                {
+                    condition = "WHERE sub_id LIKE '" + search.Text.ToString() + "%'";
+                }
+                if (settedcb == "Year")
+                {
+                    condition = "WHERE stud_year LIKE '" + search.Text.ToString() + "%'";
+                }
+                if (settedcb == "Description")
+                {
+                    condition = "WHERE sub_taken LIKE '" + search.Text.ToString() + "%'";
+                }
+
+
+                query = query + " " + condition;
+
+                MySqlCommand cmd = new MySqlCommand(query, conn);
+                MySqlDataReader dp = cmd.ExecuteReader();
+
+                while (dp.Read())
+                {
+                    enrollment_table.Rows.Add(dp[0].ToString(), dp[1].ToString(), dp[2].ToString(), dp[3].ToString(), dp[4].ToString());
+                }
+                dp.Close();
+
+
+            }
+            catch (Exception e)
+            {
+                MessageBox.Show("Error syntax: " + e.Message);
+            }
+
+        }
 
 
         //=====================================================
@@ -72,17 +132,52 @@ namespace Admin
         private void Enrollment_Load(object sender, EventArgs e)
         {
             databaseConnection();
+            displayEnrollmentTable();
+            set_CB.Text = "--Set--";
+            search.Text = "";
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
+            if (selectedId == "0")
+            {
+                MessageBox.Show("Select student to delete..", "Notice");
+            }
+            else
+            {
+                deleteValue(selectedId, selecteddesc);
+
+            }
+            displayEnrollmentTable();
+            displaySearchResult();
+     
+        }
+
+        
+        private void enrollment_table_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+            selectedId = enrollment_table.SelectedCells[0].Value.ToString();
+            selectedstudid = enrollment_table.SelectedCells[1].Value.ToString();
+            selectedyear = enrollment_table.SelectedCells[2].Value.ToString();
+            selectedsubid = enrollment_table.SelectedCells[3].Value.ToString();
+            selecteddesc = enrollment_table.SelectedCells[4].Value.ToString();
+         
+        }
+
+        private void button1_Click(object sender, EventArgs e)
+        {
 
         }
 
-        private void enrollment_table_CellClick(object sender, DataGridViewCellEventArgs e)
+        private void set_CB_SelectedIndexChanged(object sender, EventArgs e)
         {
-            string id = enrollment_table.SelectedCells[0].Value.ToString();
-            deleteSelectedRow(id);
+            settedcb = set_CB.Text;
+            displaySearchResult();
+        }
+
+        private void search_TextChanged(object sender, EventArgs e)
+        {
+            displaySearchResult();
         }
     }
 }
